+++ src/chain/mod.rs
@@ -33,6 +33,8 @@ use tokio::io::{AsyncWrite, AsyncWriteExt, BufWriter};
 #[derive(Debug, Clone, Default)]
 pub struct ExportOptions {
     pub skip_checksum: bool,
+    pub include_receipts: bool,
+    pub include_events: bool,
     pub seen: CidHashSet,
 }
 
@@ -140,8 +142,14 @@ async fn export_to_forest_car<D: Digest>(
 ) -> anyhow::Result<Option<digest::Output<D>>> {
     let ExportOptions {
         skip_checksum,
+        include_receipts,
+        include_events,
         seen,
     } = options.unwrap_or_default();
+
+    if include_events && !include_receipts {
+        anyhow::bail!("message receipts must be included when events are included");
+    }
 
     let stateroot_lookup_limit = tipset.epoch() - lookup_depth;
 
@@ -161,6 +169,8 @@ async fn export_to_forest_car<D: Digest>(
             stateroot_lookup_limit,
         )
         .with_seen(seen)
+        .with_message_receipts(include_receipts)
+        .with_events(include_events)
         .track_progress(true),
     );